services/gps_producer/requirements.txt
kafka-python
psycopg2-binary
folium